+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Fixed
 - #55 Comments at the top of a block before a declaration.
 
+## [0.11.2.1] - 2015-10-06
+### Added
+- Type qualifiers are now allowed before an antiquoted type.
+
 ## [0.11.2] - 2015-09-29
 ### Added
 - `qqPat` and `qqExp` are now exposed.